advance getting data from bookmarks from syncronization
