+++ src/components/map/Sidebar/Bookmarks.tsx
@@ -9,17 +9,32 @@ import { BookmarksList } from "./BookmarksList";
 // import "./Bookmarks.css";
 type BookmarksProps = {
   openLinkedNode: any;
+  bookmarkedUserNodes: any[];
 };
 
 const Bookmarks = (props: BookmarksProps) => {
+  console.log(" 00>> props.bookmarkedUserNodes", props.bookmarkedUserNodes);
+  // const updatedBookmarks = props.bookmarkedUserNodes.filter(cur=>cur)
   const tabsItems = [
     {
       title: "Updated",
-      content: <BookmarksList openLinkedNode={props.openLinkedNode} updates={true} /*openSection={false}*/ />,
+      content: (
+        <BookmarksList
+          openLinkedNode={props.openLinkedNode}
+          updates={true}
+          bookmarks={props.bookmarkedUserNodes} /*openSection={false}*/
+        />
+      ),
     },
     {
       title: "Studied",
-      content: <BookmarksList openLinkedNode={props.openLinkedNode} updates={false} /*openSection={true} */ />,
+      content: (
+        <BookmarksList
+          openLinkedNode={props.openLinkedNode}
+          updates={false}
+          bookmarks={props.bookmarkedUserNodes} /*openSection={true} */
+        />
+      ),
     },
   ];
